+++ AuctionExpress.Service/TransactionService.cs
@@ -106,12 +106,13 @@ namespace AuctionExpress.Service
         }
 
 
-        //Update Transaction <--Update IsPaid, currently Sellers allowed to edit
+        //Update Transaction <--Updates IsPaid, currently Sellers allowed to edit
 
         public bool UpdateTransaction(TransactionEdit model)
         {
             using (var ctx = new ApplicationDbContext())
             {
+
                 var entity =
                     ctx
                     .Transaction
@@ -126,7 +127,7 @@ namespace AuctionExpress.Service
             }
         }
 
-        public bool DeleteTransaction(int transactionId)
+        public string DeleteTransaction(int transactionId)
         {
             using (var ctx = new ApplicationDbContext())
             {
@@ -134,11 +135,25 @@ namespace AuctionExpress.Service
                     ctx
                         .Transaction
                         .Single(e => e.TransactionId == transactionId);
+                if (entity.TransactionProduct.ProductSeller == _userId.ToString())
+                {
+                    try
+                    {
+                        ctx.Transaction.Remove(entity);
+                        ctx.SaveChanges();
 
-                ctx.Transaction.Remove(entity);
+                        return "Transaction successfully deleted";
+                    }
+                    catch (Exception e)
+                    {
+                        return e.Message;
+                    }
+                }
 
-                return ctx.SaveChanges() == 1;
+                return "User unauthorized to delete this transaction.";
             }
+
         }
     }
+
 }
